restructure routes and handler, add user based on req.user
