+++ src/server/config/keys.js
@@ -3,3 +3,7 @@ export const cookieKey = 'cookieKeycookie'
 export const GoogleClientID =
     '415704818384-for77vqcd1bf73pfv8a56tmf3s5grg98.apps.googleusercontent.com'
 export const GoogleClientSecret = 'c2vPSelfPQTjbF3DRUfmqJIJ'
+
+// FOOD API
+export const NutritionXAppID = '559e441c'
+export const NutritionXAppKeys = 'b5372722af0247ff5cda75e3fdf36c2c'
