+++ src/server/handler/api/food.js
@@ -1,0 +1,3 @@
+export const handleGetFood = (req, res, next) => {
+    console.log(req.user)
+}
