+++ src/server/handler/auth.js
@@ -1,0 +1,7 @@
+export const handleAuthCallback = (req, res, next) => {
+    res.redirect('/')
+}
+
+export const handleGetCurrentUser = (req, res, next) => {
+    res.send(req.user)
+}
