+++ src/server/routes/api/food.js
@@ -1,0 +1,5 @@
+import { handleGetFood } from '@server/handler/api/food'
+
+export default function(r) {
+    r.get('/api/food', handleGetFood)
+}
