+++ src/server/routes/index.js
@@ -1,7 +1,13 @@
 import auth from './auth'
 import template from './template'
 
-export default function(app) {
+import foodAPI from './api/food'
+
+export default app => {
     auth(app)
+
+    foodAPI(app)
+
+    // Template receive * url
     template(app)
 }
